[js] change the logic how products updated, see notes below:
Replacing Elements with `$parent.innerHTML = ''` causing bug in IE11 and below, which results in Elements of `$parent` loosing their children (that's even sounding awful! IE, how could you!!).

Well, in fact, we shouldn't completely remove Elements after filtering and sorting, instead it would be enough to hide filtered items and append everything updated to the end of $parent. It works in IE9 and above.