[data][modules] deal with the fact that `matter` and `images` will not exist while tests executing
Tests will require Grunt, and it will force everything in tasks to execute at least once. Data index file is a part of Gruntfile, and thus it will be executed too. It won't be happy not to find `matter` and `images` files at this point, so we need to lazy load them through functions wrapping